refactor(actor): rename wasm to ffi + concretise ctx structs (#663 phase C)
Renames the actor SDK's binding layer from wasm to ffi to reflect that
any host satisfying the _p32 import surface (today the wasm runtime in
aether_substrate::actor::wasm; future C / OS-process hosts) can drive
an actor through it. Substrate's actor::wasm module keeps its name —
that IS the wasm runtime, one consumer of the FFI-actor contract.

Symbol renames across ~25 files:
- WasmTransport / WASM_TRANSPORT -> FfiTransport / FFI_TRANSPORT
- WasmCtx / WasmInitCtx / WasmDropCtx -> FfiCtx / FfiInitCtx / FfiDropCtx
- WasmActor -> FfiActor
- aether_actor::wasm::* -> aether_actor::ffi::*
- crate::wasm::raw -> crate::ffi::raw (inside aether-actor only)

Drops the Component back-compat alias on FfiActor — no live consumers
in tree.

Concrete ctx structs:
- FfiTransport moved to aether-actor/src/ffi/transport.rs
- FfiCtx / FfiInitCtx / FfiDropCtx are now concrete structs in
  aether-actor/src/ffi/ctx.rs, no longer aliases over the parametric
  Ctx<'a, T>. Each pins type Transport = FfiTransport in its
  MailSender / OutboundReply / Resolver / Persistence impls. Inherent
  shims (transport, actor, resolve_actor, reply_to, ...) are kept on
  the concrete structs so callers don't need to import the per-stage
  capability traits at this phase — Phase D's generic-bounds API
  introduces explicit trait bounds at the FfiActor signature.

Retires the parametric Ctx<'a, T> / InitCtx<'a, T> / DropCtx<'a, T>
core in aether-actor/src/actor/ctx/parametric.rs; the per-stage
capability traits are the cross-transport contract going forward.

Updates aether-actor-derive's #[actor] macro path emissions to point
at the renamed Ffi* types.

Refs #663.
