+++ crates/aether-actor/examples/hello.rs
@@ -16,7 +16,7 @@
 //! MCP via the same section so the harness sees typed capabilities
 //! plus author-written intent for each inbox.
 
-use aether_actor::{BootError, KindId, WasmActor, WasmCtx, WasmInitCtx, actor};
+use aether_actor::{BootError, KindId, FfiActor, FfiCtx, FfiInitCtx, actor};
 use aether_capabilities::RenderCapability;
 use aether_kinds::{DrawTriangle, Ping, Pong, Tick, Vertex};
 
@@ -63,10 +63,10 @@ pub struct Hello {
 /// `aether.ping` with an incrementing `seq` to exercise reply-to-
 /// sender; the matching `aether.pong` lands back at your session.
 #[actor]
-impl WasmActor for Hello {
+impl FfiActor for Hello {
     const NAMESPACE: &'static str = "hello";
 
-    fn init(ctx: &mut WasmInitCtx<'_>) -> Result<Self, BootError> {
+    fn init(ctx: &mut FfiInitCtx<'_>) -> Result<Self, BootError> {
         Ok(Hello {
             pong: ctx.resolve::<Pong>(),
         })
@@ -79,7 +79,7 @@ impl WasmActor for Hello {
     /// its own tick loop. The effect is visible in `capture_frame`
     /// output.
     #[handler]
-    fn on_tick(&mut self, ctx: &mut WasmCtx<'_>, _tick: Tick) {
+    fn on_tick(&mut self, ctx: &mut FfiCtx<'_>, _tick: Tick) {
         ctx.actor::<RenderCapability>().send(&TRIANGLE);
     }
 
@@ -92,7 +92,7 @@ impl WasmActor for Hello {
     /// The seq echo lets you pair requests and replies when multiple
     /// are in flight.
     #[handler]
-    fn on_ping(&mut self, ctx: &mut WasmCtx<'_>, ping: Ping) {
+    fn on_ping(&mut self, ctx: &mut FfiCtx<'_>, ping: Ping) {
         if let Some(sender) = ctx.reply_to() {
             ctx.reply(sender, self.pong, &Pong { seq: ping.seq });
         }
