+++ crates/aether-actor/src/actor/ctx/mod.rs
@@ -1,33 +1,32 @@
-//! Per-stage capability traits + the parametric ctx structs that back
-//! the FFI guest path today.
+//! Per-stage capability traits — the cross-transport ctx contract.
 //!
-//! Issue 663 phase A factors the ctx surface into per-stage trait
-//! files — each describes a slice of functionality applicable at one
+//! Each trait describes a slice of functionality applicable at one
 //! lifecycle stage:
 //!
 //! - [`MailSender`] — outbound mail (every ctx).
 //! - [`OutboundReply`] — reply-to-originator (per-handler ctxs only).
 //! - [`Resolver`] — init-time mailbox/kind resolution (init ctxs only).
-//! - [`Persistence`] — `replace_component` migration bundle (drop ctxs only).
-//! - [`LifecycleControl`] — self-shutdown + monitor (per-handler ctxs that
-//!   participate in ADR-0079 lifecycle).
+//! - [`Persistence`] — `replace_component` migration bundle (drop
+//!   ctxs only).
+//! - [`LifecycleControl`] — self-shutdown + monitor (per-handler ctxs
+//!   that participate in ADR-0079 lifecycle).
 //!
-//! The parametric [`Ctx`] / [`InitCtx`] / [`DropCtx`] structs in
-//! [`parametric`] back the FFI guest path today; phase C concretises
-//! them into [`crate::wasm::WasmCtx`] / etc. (renamed to `FfiCtx` in
-//! the same phase) and retires the parametric core. The trait surface
-//! defined here will remain the user-facing contract.
+//! The concrete ctx structs live next to their transport: FFI-side
+//! `FfiInitCtx` / `FfiCtx` / `FfiDropCtx` in [`crate::ffi::ctx`];
+//! native-side `NativeInitCtx` / `NativeCtx` in
+//! `aether_substrate::actor::native::ctx`. Each impls the trait
+//! subset applicable to its stage; default-impl bodies on
+//! [`MailSender`] cover the routing methods so the per-impl code is
+//! the stage-specific accessors.
 
 pub mod lifecycle;
 pub mod mail_sender;
 pub mod outbound_reply;
-pub mod parametric;
 pub mod persistence;
 pub mod resolver;
 
 pub use lifecycle::LifecycleControl;
 pub use mail_sender::MailSender;
 pub use outbound_reply::OutboundReply;
-pub use parametric::{Ctx, DropCtx, InitCtx};
 pub use persistence::Persistence;
 pub use resolver::Resolver;
