+++ crates/aether-camera/src/runtime.rs
@@ -35,7 +35,7 @@
 
 use std::collections::HashMap;
 
-use aether_actor::{BootError, WasmActor, WasmCtx, WasmInitCtx, actor};
+use aether_actor::{BootError, FfiActor, FfiCtx, FfiInitCtx, actor};
 use aether_capabilities::RenderCapability;
 use aether_kinds::{Camera, Tick, WindowSize};
 use aether_math::{Mat4, PI, Quat, Vec2, Vec3};
@@ -246,10 +246,10 @@ pub struct CameraComponent {
 ///
 /// Use `capture_frame` between sends to verify each change.
 #[actor]
-impl WasmActor for CameraComponent {
+impl FfiActor for CameraComponent {
     const NAMESPACE: &'static str = "camera";
 
-    fn init(_ctx: &mut WasmInitCtx<'_>) -> Result<Self, BootError> {
+    fn init(_ctx: &mut FfiInitCtx<'_>) -> Result<Self, BootError> {
         let mut cameras = HashMap::new();
         cameras.insert(
             "main".to_owned(),
@@ -272,7 +272,7 @@ impl WasmActor for CameraComponent {
     /// # Agent
     /// Tick-driven; not useful to send manually.
     #[handler]
-    fn on_tick(&mut self, ctx: &mut WasmCtx<'_>, _tick: Tick) {
+    fn on_tick(&mut self, ctx: &mut FfiCtx<'_>, _tick: Tick) {
         for cam in self.cameras.values_mut() {
             cam.mode.tick();
         }
@@ -291,7 +291,7 @@ impl WasmActor for CameraComponent {
     /// Publish-subscribe; the substrate pulses this every tick. Not
     /// useful to send manually.
     #[handler]
-    fn on_window_size(&mut self, _ctx: &mut WasmCtx<'_>, size: WindowSize) {
+    fn on_window_size(&mut self, _ctx: &mut FfiCtx<'_>, size: WindowSize) {
         if size.width > 0 && size.height > 0 {
             self.aspect = size.width as f32 / size.height as f32;
         }
@@ -302,7 +302,7 @@ impl WasmActor for CameraComponent {
     /// camera instead. Newly-created cameras are not made active
     /// automatically; pair with `set_active` to switch publishing.
     #[handler]
-    fn on_create(&mut self, _ctx: &mut WasmCtx<'_>, msg: CameraCreate) {
+    fn on_create(&mut self, _ctx: &mut FfiCtx<'_>, msg: CameraCreate) {
         if self.cameras.contains_key(&msg.name) {
             tracing::warn!(
                 target: "aether_camera",
@@ -324,7 +324,7 @@ impl WasmActor for CameraComponent {
     /// publishing pauses (no `aether.camera` mail goes out) until
     /// `set_active` picks a survivor.
     #[handler]
-    fn on_destroy(&mut self, _ctx: &mut WasmCtx<'_>, msg: CameraDestroy) {
+    fn on_destroy(&mut self, _ctx: &mut FfiCtx<'_>, msg: CameraDestroy) {
         self.cameras.remove(&msg.name);
         if self.active.as_deref() == Some(msg.name.as_str()) {
             self.active = None;
@@ -335,7 +335,7 @@ impl WasmActor for CameraComponent {
     /// `"aether.render"` each tick. Errors (warn-log, no state
     /// change) if `name` isn't bound.
     #[handler]
-    fn on_set_active(&mut self, _ctx: &mut WasmCtx<'_>, msg: CameraSetActive) {
+    fn on_set_active(&mut self, _ctx: &mut FfiCtx<'_>, msg: CameraSetActive) {
         if self.cameras.contains_key(&msg.name) {
             self.active = Some(msg.name);
         } else {
@@ -352,7 +352,7 @@ impl WasmActor for CameraComponent {
     /// per-mode compiled defaults. No-op (warn-log) if `name` isn't
     /// bound.
     #[handler]
-    fn on_set_mode(&mut self, _ctx: &mut WasmCtx<'_>, msg: CameraSetMode) {
+    fn on_set_mode(&mut self, _ctx: &mut FfiCtx<'_>, msg: CameraSetMode) {
         match self.cameras.get_mut(&msg.name) {
             Some(cam) => cam.mode = ModeState::from_init(&msg.mode),
             None => tracing::warn!(
@@ -368,7 +368,7 @@ impl WasmActor for CameraComponent {
     /// (warn-log) if the camera doesn't exist or is in a different
     /// mode.
     #[handler]
-    fn on_orbit_set(&mut self, _ctx: &mut WasmCtx<'_>, msg: CameraOrbitSet) {
+    fn on_orbit_set(&mut self, _ctx: &mut FfiCtx<'_>, msg: CameraOrbitSet) {
         match self.cameras.get_mut(&msg.name) {
             Some(cam) => match &mut cam.mode {
                 ModeState::Orbit(state) => state.apply(&msg.params),
@@ -391,7 +391,7 @@ impl WasmActor for CameraComponent {
     /// semantics as `orbit.set` for the orthographic mode's `center`
     /// / `extent`.
     #[handler]
-    fn on_topdown_set(&mut self, _ctx: &mut WasmCtx<'_>, msg: CameraTopdownSet) {
+    fn on_topdown_set(&mut self, _ctx: &mut FfiCtx<'_>, msg: CameraTopdownSet) {
         match self.cameras.get_mut(&msg.name) {
             Some(cam) => match &mut cam.mode {
                 ModeState::Topdown(state) => state.apply(&msg.params),
