+++ crates/aether-substrate/src/actor/wasm/mod.rs
@@ -1,5 +1,5 @@
 //! Wasm-flavoured actor primitives: the [`Component`] trait and per-mail
-//! [`ComponentCtx`] (the `aether-actor::WasmActor` counterpart that the
+//! [`ComponentCtx`] (the `aether-actor::FfiActor` counterpart that the
 //! substrate's wasmtime trampoline drives), the linker registration for
 //! the host fns the guest imports, and the wasm-only support tables
 //! ([`reply_table`] for in-flight `wait_reply` correlations,
